update form component
